[Feat] finish signUp tabs and submit limit
