+++ src/common/redux/store.ts
@@ -1,11 +1,13 @@
 import { configureStore, combineReducers } from '@reduxjs/toolkit';
 import userCenterSlice from './feature/userCenter';
-import loginTabs from './feature/loginTabs';
+import loginTabsSlice from './feature/loginTabs';
+import signUpSlice from './feature/signUp';
 import { login } from './service/login';
 
 const reducers = combineReducers({
   userCenterSlice,
-  loginTabs,
+  signUpSlice,
+  loginTabsSlice,
   [login.reducerPath]: login.reducer,
 });
